requirements.txt
flask==3.0.0
openai==1.2.3
python-dotenv
